Updated scripts to use capture instead of remember
As of org-mode version 8, remember has been completely replained with
capture. The old code, designed to work with capture, no longer works
with org-mode. This requires a minro change to the remember script and a
small change to the emacs configuration.

This patch is based on the default capture template configuration. As a
result, the behavior from this patch is very similar, but  not
identical. The new capture templates can be tweaked.